Backend completado: API y base de datos configuradas

+++ src/main/java/com/codepair/petadopt/repository/SeguimientoRepository.java
@@ -1,4 +1,3 @@
-// SeguimientoRepository.java
 package com.codepair.petadopt.repository;
 
 import com.codepair.petadopt.model.Seguimiento;
@@ -6,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SeguimientoRepository extends JpaRepository<Seguimiento, Integer> {
+public interface SeguimientoRepository extends JpaRepository<Seguimiento, Long> {
 }
