+++ src/main/java/com/codepair/petadopt/repository/SolicitudAdopcionRepository.java
@@ -1,4 +1,3 @@
-// SolicitudAdopcionRepository.java
 package com.codepair.petadopt.repository;
 
 import com.codepair.petadopt.model.SolicitudAdopcion;
@@ -6,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SolicitudAdopcionRepository extends JpaRepository<SolicitudAdopcion, Integer> {
+public interface SolicitudAdopcionRepository extends JpaRepository<SolicitudAdopcion, Long> {
 }
