+++ src/main/java/com/codepair/petadopt/repository/UsuarioRepository.java
@@ -1,4 +1,3 @@
-// UsuarioRepository.java
 package com.codepair.petadopt.repository;
 
 import com.codepair.petadopt.model.Usuario;
@@ -6,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 }
